Finished homework
The assignment was very relevant to what we have covered in class. Plus, it was a great practice of the use of the documntations such as W3Schools.

I have faced some difficulties in trying to use the same CSS file for both HTML pages. It was somewhat confusing. I think there might be a better practice in which we can comfortably use the same CSS for more than one HTML files without having to make so many changes. I would like to know more about that.